# pkg/web/static/openapi.yaml
openapi: 3.0.3
info:
  title: TRISA Node API
  description: >-
    The TRISA Node API allows users to interact with the TRISA Self-Hosted Node
    in a programmatic fashion.

    It is intended that this API is served on an internal IP address with
    routing only available to internal services.
  termsOfService: https://trisa.io/terms/
  contact:
    email: support@rotational.io
  license:
    name: MIT Licensed
    url: https://github.com/trisacrypto/trisa/blob/main/LICENSE
  version: v0.23.0
...
paths:
...
  /v1/counterparties:
    get:
      tags:
        - counterparty
      summary: List counterparties
      description: Paginated list of all stored counterparties
      operationId: listCounterparties
      security:
        - bearer-auth: [ ]
      parameters:
        - name: page_size
          in: query
          schema:
            type: integer
        - name: next_page_token
          in: query
          schema:
            type: string
        - name: prev_page_token
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CounterpartyList'
    post:
      tags:
        - counterparty
      summary: Create counterparty
      description: Create a new counterparty
      operationId: createCounterparty
      security:
        - bearer-auth: [ ]
      requestBody:
        description: Create a new counterparty
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Counterparty'
        required: true
      responses:
        '200':
          description: Counterparty created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Counterparty'
        '400':
          description: Invalid input
        '422':
          description: Validation exception or missing field
...
  /v1/counterparties/{counterpartyID}:
    get:
      tags:
        - counterparty
      summary: Find counterparty by ID
      description: Returns a single counterparty if found
      operationId: counterpartyDetail
      security:
        - bearer-auth: [ ]
      parameters:
        - name: counterpartyID
          in: path
          description: ID of counterparty to return
          required: true
          schema:
            type: string
            format: ULID
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Counterparty'
        '404':
          description: Counterparty not found
    put:
      tags:
        - counterparty
      summary: Updates a counterparty record
      description: Update a counterparty record (does not patch, all fields are required)
      operationId: updateCounterparty
      security:
        - bearer-auth: [ ]
      parameters:
        - name: counterpartyID
          in: path
          description: ID of counterparty to update
          required: true
          schema:
            type: string
            format: ULID
      requestBody:
        description: Updated counterparty record
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Counterparty'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Counterparty'
        '400':
          description: Invalid input
        '422':
          description: Validation exception or missing field
    delete:
      tags:
        - counterparty
      summary: Deletes a counterparty
      description: Deletes a counterparty
      operationId: deleteCounterparty
      security:
        - bearer-auth: [ ]
      parameters:
        - name: counterpartyID
          in: path
          description: ID of counterparty to delete
          required: true
          schema:
            type: string
            format: ULID
      responses:
        '200':
          description: Successful operation
        '404':
          description: Counterparty not found
...
components:
  schemas:
...
    PageQuery:
      type: object
      description: The request object for paginated endpoints.
      properties:
        page_size:
          type: integer
          example: 50
          description: The number of items to return per page.
        next_page_token:
          type: string
          format: base64
          description: The token for retrieving the next page of results.
        prev_page_token:
          type: string
          format: base64
          description: The token for retrieving the previous page of results.
...
    Counterparty:
      type: object
      description: The counterparty object.
      properties:
        id:
          type: string
          format: ULID
          description: The unique identifier of the counterparty.
        source:
          type: string
          description: The source of the counterparty.
        directory_id:
          type: string
          description: The unique identifier of the directory that contains the counterparty.
        registered_directory:
          type: string
          description: The name of the registered directory that contains the counterparty.
        protocol:
          type: string
          description: The protocol used to communicate with the counterparty.
        common_name:
          type: string
          description: The common name of the counterparty.
        endpoint:
          type: string
          description: The endpoint URL of the counterparty.
        travel_address:
          type: string
          description: The travel address of the counterparty.
        name:
          type: string
          description: The name of the counterparty.
        website:
          type: string
          description: The website URL of the counterparty.
        country:
          type: string
          description: The country of the counterparty.
        business_category:
          type: string
          description: The business category of the counterparty.
        vasp_categories:
          type: array
          items:
            type: string
          description: The list of VASP categories that the counterparty belongs to.
        verified_on:
          type: string
          format: date-time
          description: The date and time when the counterparty was verified.
        ivms101:
          type: string
          description: The IVMS101 certificate of the counterparty.
        created:
          type: string
          format: date-time
          description: The date and time when the counterparty was created.
        modified:
          type: string
          format: date-time
          description: The date and time when the counterparty was last modified.
    CounterpartyList:
      type: object
      description: The response object for the counterparties endpoint.
      properties:
        page:
          $ref: '#/components/schemas/PageQuery'
        counterparties:
          type: array
          items:
            $ref: '#/components/schemas/Counterparty'
...
  securitySchemes:
    bearer-auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
servers:
  - url: http://localhost:8000
    variables: { }
    description: Local Development Server